Use patched LayerMapping to allow updating reports
